+++ client/src/main/java/Network/NetworkManager.java
@@ -4,8 +4,8 @@ import Messages.Channels;
 import Messages.MoveInfo;
 import Messages.PlayerInfo;
 import Messages.RoomInfo;
-import Network.Thread.TaskThread;
-import Network.Thread.TaskThreadFactory;
+import TaskThreads.TaskThread;
+import TaskThreads.TaskThreadFactory;
 import com.pubnub.api.PNConfiguration;
 import com.pubnub.api.PubNub;
 import com.pubnub.api.callbacks.PNCallback;
